added comments/documentation to all necessary classes

+++ src/main/java/com/account/manager/am/dao/TransactionRepository.java
@@ -3,5 +3,5 @@ package com.account.manager.am.dao;
 import com.account.manager.am.model.Transaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionRepository extends JpaRepository<Transaction, Long> {
+public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
 }
